requirements.txt
numpy>=1.26
pandas>=2.2
polars>=0.19
pyarrow>=16.1
scikit-learn>=1.5
torch>=2.2
tqdm>=4.66
pydantic>=2.7
tensorboard>=2.14
pyyaml>=6.0
fastapi>=0.115
uvicorn>=0.30
streamlit>=1.38
matplotlib>=3.8
umap-learn>=0.5
kaggle>=1.5
boto3>=1.26
